requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
tensorflow==2.15.0
pillow==10.0.1
numpy==1.25.2